<fix>修复了坐标系和约束赋值的一部分缺陷

+++ src/component/constrained/single_point_constrained.cpp
@@ -32,7 +32,7 @@ void SinglePointConstrains::SetComp(const file_data &datas) {
     SPC temp_spc;
     temp_spc._node = boost::any_cast<int>(card.at(2 + 3 * i));
     temp_spc._component = boost::any_cast<int>(card.at(3 + 3 * i));
-    temp_spc._value = boost::any_cast<int>(card.at(4 + 3 * i));
+    temp_spc._value = boost::any_cast<double>(card.at(4 + 3 * i));
     _spc.push_back(temp_spc);
   }
 }
